+++ tests/issue_59.spec.ts
@@ -1,4 +1,5 @@
-import { memoize } from '../src/index';
+import { describe, expect, it } from 'vitest';
+import { memoize } from 'proxy-memoize';
 
 describe('issue #59', () => {
   it('no duplicated property access', () => {
@@ -8,14 +9,17 @@ describe('issue #59', () => {
     const getA = memoize((state: State) => state.a);
 
     let count = 0;
-    const state2: State = new Proxy({ a: 11, b: 22 }, {
-      get(target, prop: 'a' | 'b') {
-        if (prop === 'a') {
-          count += 1;
-        }
-        return target[prop];
+    const state2: State = new Proxy(
+      { a: 11, b: 22 },
+      {
+        get(target, prop: 'a' | 'b') {
+          if (prop === 'a') {
+            count += 1;
+          }
+          return target[prop];
+        },
       },
-    });
+    );
 
     expect(getA(state1)).toBe(1);
     expect(getA(state2)).toBe(11);